requirements.txt
bigquery = 2.26.0
googleads = 13.0.0